Quick fix: avoid losing node clicks
Problem: User clicks on nodes could get lost if new *messages* were
created between the click and the processing of the event. This is
probably because messages were put below nodes on the DOM, messing up
identifiers.

Solution: Paint messages on top of nodes. This could be perceived as
aesthetically less pleasing but right now I find it more pleasing to be
able to click on a node many times that it bothers me that messages fly
over nodes until reaching their center.